pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "structured-search"
version = "0.1.0"
description = "Structured, auditable LLM workflows with a unified CLI: precision, traceability, reproducible flows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.0",
  "pyyaml",
  "jsonschema>=4.0.0",
]

[project.optional-dependencies]
ollama = ["langchain-ollama>=0.2"]
api = ["fastapi>=0.115", "uvicorn[standard]>=0.30"]

[project.scripts]
structured-search = "structured_search.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/structured_search"]

[dependency-groups]
dev = [
    "import-linter>=2.4",
    "pre-commit>=4.0",
    "pytest>=9.0.2",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = [
  "E501",    # line-length handled by formatter
  "UP007",   # X | Y union syntax (we use it intentionally with from __future__)
  "UP046",   # Generic[T] -> type params (deferred: requires broader refactor)
]

[tool.ruff.lint.isort]
known-first-party = ["structured_search"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["unit_test_*.py", "integration_test_*.py", "api_test_*.py", "test_*.py", "*_test.py"]
markers = [
    "integration: integration tests that require real config files (atoms, JSONL)",
]
